Fix encoder parameters and launch files
Fixed encoder resolution, wheel diameter, and wheel separation
to match actual robot. Fixed tabletop offsets to match improved
odometry. Copied PRSG2 configuration to this package so that it
can be configured for the robot without interfering with other
users of the prsg2_robot package. Tightened tolerances for
motion vectors launch, and decreased frame rate to 20 fps.